Title: 자연수 n개의 합, Time: 84ms, Memory: 8MB, Status: Wrong Answer - CodeTree

+++ 240909/자연수 n개의 합/sum-of-n-natural-numbers.java
@@ -29,7 +29,7 @@ public class Main {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer st;
 
-        s = Integer.parseInt(br.readLine());
+        s = Long.parseLong(br.readLine());
 
         long ans = binarySearch(s);
